Split CrawlerLogic from CrawlerOverviewLogic (#110850)
