+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/crawler/crawler_router.test.tsx
@@ -4,6 +4,9 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+import { setMockActions } from '../../../__mocks__/kea_logic';
+import '../../../__mocks__/shallow_useeffect.mock';
+import '../../__mocks__/engine_logic.mock';
 
 import React from 'react';
 
@@ -14,14 +17,25 @@ import { CrawlerRouter } from './crawler_router';
 import { CrawlerSingleDomain } from './crawler_single_domain';
 
 describe('CrawlerRouter', () => {
+  const mockActions = {
+    fetchCrawlerData: jest.fn(),
+    getLatestCrawlRequests: jest.fn(),
+  };
+
   let wrapper: ShallowWrapper;
 
   beforeEach(() => {
     jest.clearAllMocks();
+    setMockActions(mockActions);
     wrapper = shallow(<CrawlerRouter />);
   });
 
-  it('renders a crawler single domain view', () => {
+  it('calls fetchCrawlerData and starts polling on page load', () => {
+    expect(mockActions.fetchCrawlerData).toHaveBeenCalledTimes(1);
+    expect(mockActions.getLatestCrawlRequests).toHaveBeenCalledWith(false);
+  });
+
+  it('renders a crawler views', () => {
     expect(wrapper.find(CrawlerOverview)).toHaveLength(1);
     expect(wrapper.find(CrawlerSingleDomain)).toHaveLength(1);
   });
