+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/crawler/crawler_router.tsx
@@ -5,15 +5,26 @@
  * 2.0.
  */
 
-import React from 'react';
+import React, { useEffect } from 'react';
 import { Route, Switch } from 'react-router-dom';
 
+import { useActions } from 'kea';
+
 import { ENGINE_CRAWLER_DOMAIN_PATH, ENGINE_CRAWLER_PATH } from '../../routes';
+
+import { CrawlerLogic } from './crawler_logic';
 
 import { CrawlerOverview } from './crawler_overview';
 import { CrawlerSingleDomain } from './crawler_single_domain';
 
 export const CrawlerRouter: React.FC = () => {
+  const { fetchCrawlerData, getLatestCrawlRequests } = useActions(CrawlerLogic);
+
+  useEffect(() => {
+    fetchCrawlerData();
+    getLatestCrawlRequests(false);
+  }, []);
+
   return (
     <Switch>
       <Route exact path={ENGINE_CRAWLER_PATH}>
